produto service layer validations

+++ src/main/java/com/example/demo/domain/services/impl/ProdutoDomainServiceImpl.java
@@ -9,10 +9,13 @@ import org.springframework.stereotype.Service;
 
 import com.example.demo.application.dtos.ProdutoRequestDto;
 import com.example.demo.application.dtos.ProdutoResponseDto;
+import com.example.demo.domain.exceptions.ProdutoComNomeDuplicadoException;
 import com.example.demo.domain.models.entities.Produto;
 import com.example.demo.domain.services.interfaces.ProdutoDomainService;
 import com.example.demo.infrastructure.repositories.CategoriaRepository;
 import com.example.demo.infrastructure.repositories.ProdutoRepository;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class ProdutoDomainServiceImpl implements ProdutoDomainService {
@@ -29,6 +32,12 @@ public class ProdutoDomainServiceImpl implements ProdutoDomainService {
 	@Override
 	public String registrarProduto(ProdutoRequestDto request) {
 
+		if (produtoRepository.existsByNome(request.getNome()))
+			throw new ProdutoComNomeDuplicadoException(request.getNome());
+
+		if (!categoriaRepository.existsById(request.getCategoriaId()))
+			throw new EntityNotFoundException("Categoria com ID " + request.getCategoriaId() + " não encontrada.");
+
 		var produto = modelMapper.map(request, Produto.class);
 		produto.setId(UUID.randomUUID());
 		produto.setCategoria(categoriaRepository.findById(request.getCategoriaId()).get());
@@ -40,8 +49,15 @@ public class ProdutoDomainServiceImpl implements ProdutoDomainService {
 
 	@Override
 	public ProdutoResponseDto atualizarProduto(UUID id, ProdutoRequestDto request) {
+		
+		if (produtoRepository.existsByNome(request.getNome()))
+			throw new ProdutoComNomeDuplicadoException(request.getNome());
 
-		var produto = produtoRepository.findById(id).get();
+		var produto = produtoRepository.findById(id)
+				.orElseThrow(() -> new EntityNotFoundException("Produto com ID " + id + " não encontrada."));
+
+		if (!categoriaRepository.existsById(request.getCategoriaId()))
+			throw new EntityNotFoundException("Categoria com ID " + request.getCategoriaId() + " não encontrada.");
 
 		produto.setNome(request.getNome());
 		produto.setPreco(request.getPreco());
@@ -56,6 +72,9 @@ public class ProdutoDomainServiceImpl implements ProdutoDomainService {
 	@Override
 	public String excluirProduto(UUID id) {
 
+		if (!produtoRepository.existsById(id))
+			throw new EntityNotFoundException("Produto com ID " + id + " não encontrada.");
+
 		produtoRepository.deleteById(id);
 
 		return "Produto excluído com sucesso!";
@@ -63,6 +82,10 @@ public class ProdutoDomainServiceImpl implements ProdutoDomainService {
 
 	@Override
 	public ProdutoResponseDto consultarProdutoPorId(UUID id) {
+
+		if (!produtoRepository.existsById(id))
+			throw new EntityNotFoundException("Produto com ID " + id + " não encontrada.");
+
 		return modelMapper.map(produtoRepository.findById(id), ProdutoResponseDto.class);
 	}
